# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AlfaUnipac GV — Sistema de Gestão Acadêmica API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /retencao:
    get:
      operationId: listRetencao
      tags: [retencao]
      summary: List students in retention
      security:
        - bearerAuth: []
      parameters:
        - name: curso
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: disciplinaId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of retention records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RetencaoDetalhada"
  /retencao/{id}:
    patch:
      operationId: updateRetencao
      tags: [retencao]
      summary: Update retention record status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRetencaoBody"
      responses:
        "200":
          description: Updated retention record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Retencao"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /retencao/{id}/timeline:
    get:
      operationId: getRetencaoTimeline
      tags: [retencao]
      summary: Get action timeline for a retention record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline of actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RetencaoAuditLog"
  /retencao/{id}/notificar:
    post:
      operationId: notificarAluno
      tags: [retencao]
      summary: Notify student about retention
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificarAlunoBody"
      responses:
        "200":
          description: Notification registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetencaoAuditLog"

  # DOCUMENTOS
...
  /relatorios/faltas-por-aluno:
    get:
      operationId: getRelatorioFaltasPorAluno
      tags: [relatorios]
      summary: Report of absences by student
      security:
        - bearerAuth: []
      parameters:
        - name: dataInicio
          in: query
          schema:
            type: string
        - name: dataFim
          in: query
          schema:
            type: string
        - name: curso
          in: query
          schema:
            type: string
        - name: alunoId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelatorioFaltasAluno"
  /relatorios/faltas-por-disciplina:
    get:
      operationId: getRelatorioFaltasPorDisciplina
      tags: [relatorios]
      summary: Report of absences by discipline
      security:
        - bearerAuth: []
      parameters:
        - name: dataInicio
          in: query
          schema:
            type: string
        - name: dataFim
          in: query
          schema:
            type: string
        - name: curso
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelatorioFaltasDisciplina"
  /relatorios/retencao:
    get:
      operationId: getRelatorioRetencao
      tags: [relatorios]
      summary: Retention report
      security:
        - bearerAuth: []
      parameters:
        - name: curso
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelatorioRetencao"
  /relatorios/documentos:
    get:
      operationId: getRelatorioDocumentos
      tags: [relatorios]
      summary: Document submission report
      security:
        - bearerAuth: []
      parameters:
        - name: dataInicio
          in: query
          schema:
            type: string
        - name: dataFim
          in: query
          schema:
            type: string
        - name: tipo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelatorioDocumento"
  /relatorios/resumo-mensal:
    get:
      operationId: getRelatorioResumoMensal
      tags: [relatorios]
      summary: Monthly summary per course
      security:
        - bearerAuth: []
      parameters:
        - name: mes
          in: query
          schema:
            type: integer
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResumoMensalCurso"

  # DASHBOARD
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Retencao:
      type: object
      properties:
        id:
          type: integer
        alunoId:
          type: integer
        turmaId:
          type: integer
        percentualFaltas:
          type: number
        dataNotificacao:
          type: ["string", "null"]
        status:
          type: string
          enum: [Em_Acompanhamento, Regularizado, Reprovado_Faltas]
        observacaoSecretaria:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - alunoId
        - turmaId
        - percentualFaltas
        - status
        - createdAt
    RetencaoDetalhada:
      type: object
      properties:
        id:
          type: integer
        alunoId:
          type: integer
        turmaId:
          type: integer
        percentualFaltas:
          type: number
        dataNotificacao:
          type: ["string", "null"]
        status:
          type: string
          enum: [Em_Acompanhamento, Regularizado, Reprovado_Faltas]
        observacaoSecretaria:
          type: ["string", "null"]
        createdAt:
          type: string
        alunoNome:
          type: string
        alunoCurso:
          type: string
        disciplinaNome:
          type: string
        periodo:
          type: string
      required:
        - id
        - alunoId
        - turmaId
        - percentualFaltas
        - status
        - createdAt
        - alunoNome
        - alunoCurso
        - disciplinaNome
        - periodo
    UpdateRetencaoBody:
      type: object
      properties:
        status:
          type: string
          enum: [Em_Acompanhamento, Regularizado, Reprovado_Faltas]
        observacaoSecretaria:
          type: string
    NotificarAlunoBody:
      type: object
      properties:
        observacao:
          type: string
      required:
        - observacao
    RetencaoAuditLog:
      type: object
      properties:
        id:
          type: integer
        retencaoId:
          type: integer
        acao:
          type: string
        observacao:
          type: ["string", "null"]
        realizadoPor:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - retencaoId
        - acao
        - createdAt

    # Documento
...
    RelatorioFaltasAluno:
      type: object
      properties:
        alunoId:
          type: integer
        alunoNome:
          type: string
        curso:
          type: string
        turmaId:
          type: integer
        disciplinaNome:
          type: string
        totalAulas:
          type: integer
        presencas:
          type: integer
        faltas:
          type: integer
        faltasJustificadas:
          type: integer
        percentualFaltas:
          type: number
        emRetencao:
          type: boolean
      required:
        - alunoId
        - alunoNome
        - curso
        - turmaId
        - disciplinaNome
        - totalAulas
        - presencas
        - faltas
        - faltasJustificadas
        - percentualFaltas
        - emRetencao
    RelatorioFaltasDisciplina:
      type: object
      properties:
        disciplinaId:
          type: integer
        disciplinaNome:
          type: string
        curso:
          type: string
        turmaId:
          type: integer
        periodo:
          type: string
        totalAlunos:
          type: integer
        mediaPresenca:
          type: number
        alunosEmRetencao:
          type: integer
      required:
        - disciplinaId
        - disciplinaNome
        - curso
        - turmaId
        - periodo
        - totalAlunos
        - mediaPresenca
        - alunosEmRetencao
    RelatorioRetencao:
      type: object
      properties:
        alunoId:
          type: integer
        alunoNome:
          type: string
        curso:
          type: string
        disciplinaNome:
          type: string
        percentualFaltas:
          type: number
        status:
          type: string
        dataNotificacao:
          type: ["string", "null"]
      required:
        - alunoId
        - alunoNome
        - curso
        - disciplinaNome
        - percentualFaltas
        - status
    RelatorioDocumento:
      type: object
      properties:
        documentoId:
          type: integer
        alunoNome:
          type: string
        curso:
          type: string
        tipo:
          type: string
        dataEntrega:
          type: string
        periodo:
          type: string
        status:
          type: string
      required:
        - documentoId
        - alunoNome
        - curso
        - tipo
        - dataEntrega
        - periodo
        - status
    ResumoMensalCurso:
      type: object
      properties:
        curso:
          type: string
        totalAlunos:
          type: integer
        mediaPresenca:
          type: number
        alunosEmRetencao:
          type: integer
        totalAulas:
          type: integer
      required:
        - curso
        - totalAlunos
        - mediaPresenca
        - alunosEmRetencao
        - totalAulas

    # Dashboard
...